+++ src/main/java/nextstep/app/ui/MemberController.java
@@ -21,29 +21,9 @@ public class MemberController {
         this.memberService = memberService;
     }
 
-    public static final String AUTHORIZATION = "Basic";
-    public static final String SPLIT_CHAR = ":";
-
     @GetMapping("/members")
     public ResponseEntity<List<Member>> list(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorization) {
-        if(!authorization.startsWith(AUTHORIZATION)) {
-            throw new AuthenticationException();
-        }
-        String decodeTarget = authorization.substring(AUTHORIZATION.length()).trim();
-        String decodeValue = new String(Base64Utils.decode(decodeTarget.getBytes()));
-
-        String[] emailAndPassword = decodeValue.split(SPLIT_CHAR);
-        String email = emailAndPassword[0];
-        String password = emailAndPassword[1];
-
-        memberService.validateMember(email, password);
-
         List<Member> members = memberService.getMembers();
         return ResponseEntity.ok(members);
-    }
-
-    @ExceptionHandler(AuthenticationException.class)
-    public ResponseEntity<Void> handleAuthenticationException() {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
